# docs/swagger.yaml
basePath: /
definitions:
  main.Message:
    properties:
      messasge:
        type: string
    type: object
  main.Todo:
    properties:
      id:
        type: string
      task:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server todo server. You can visit the GitHub repository
    at https://github.com/LordGhostX/swag-gin-demo
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Todo API
  version: "1.0"
paths:
  /todo:
    get:
      operationId: get-all-todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
      summary: get all items in the todo list
    post:
      operationId: create-todo
      parameters:
      - description: Todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Message'
      summary: add a new item to the todo list
  /todo/{id}:
    delete:
      operationId: delete-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Message'
      summary: delete a todo item by ID
    get:
      operationId: get-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Message'
      summary: get a todo item by ID
    patch:
      operationId: update-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Message'
      summary: update a todo item by ID
swagger: "2.0"